refactor: create package testing-utils
