+++ packages/testing/testing-helpers/README.md
@@ -1,0 +1,3 @@
+# Testing Utils
+
+Client side utilities for running tests in the Bolt Design System. Server side utils are in `@bolt/testing-utils`.
