+++ packages/testing/testing-utils/README.md
@@ -1,0 +1,3 @@
+# Testing Utils
+
+Server side utilities for running tests in the Bolt Design System. Client side utils are in `@bolt/testing-helpers`.
