+++ packages/testing/testing-utils/__tests__/test-utils.test.js
@@ -1,7 +1,7 @@
 const { join } = require('path');
-const tu = require('../../scripts/utils/test-utils');
+const tu = require('../test-utils');
 
-const repoRoot = join(__dirname, '../..');
+const repoRoot = join(__dirname, '../../../..');
 
 describe('test-utils', () => {
   test('getPkgNameFromPath', () => {
