+++ packages/testing/testing-utils/package.json
@@ -1,0 +1,31 @@
+{
+  "name": "@bolt/testing-utils",
+  "version": "2.5.2",
+  "description": "Server side utilities for running tests in the Bolt Design System.",
+  "main": "index.js",
+  "scripts": {
+    "test:types": "tsc"
+  },
+  "dependencies": {
+    "find-pkg": "^2.0.0",
+    "fs-extra": "^8.1.0",
+    "globby": "^9.2.0"
+  },
+  "keywords": [
+    "bolt design system",
+    "design system",
+    "testing"
+  ],
+  "homepage": "https://boltdesignsystem.com",
+  "bugs": "https://github.com/bolt-design-system/bolt/issues",
+  "repository": "https://github.com/bolt-design-system/bolt/tree/master/packages/testing-utils",
+  "license": "MIT",
+  "publishConfig": {
+    "access": "public"
+  },
+  "maintainers": [{
+    "name": "Salem Ghoweri",
+    "email": "me@salemghoweri.com",
+    "web": "https://github.com/sghoweri"
+  }]
+}
